+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.1 Under development
 
+- Bug #47: Complete the Yii response lifecycle after RoadRunner emits PSR responses and update `yii2-extensions/psr-bridge` to `0.4` (@terabytesoftw)
+
 ## 0.3.0 February 28, 2026
 
 - Bug #45: Move worker resolution to `RoadRunner::run()` method and remove application container bootstrapping from the runner loop to keep application initialization inside `Application` class (@terabytesoftw)